Use local or external TCP cache, but not both.
This commit strictly divides the TCP cache methods to stop storing a local copy of the remote cache when a persistent cache exists elsewhere.

It also adds expirations to the local cache, in the event that the PHP request is allowed to process and hang for longer than the servers have collectively to be checked.