Write Queries: support "FOR UPDATE"
This change adds `FOR\s+UPDATE` to the regular expression check that is used to determine if a query is trying to write to a database.

This better supports queries from other plugins (like Woo or Wordfence) that rely on this MySQL feature.

See also: https://github.com/Automattic/HyperDB/pull/112#issuecomment-1586519047

Fixes #127.